Add isolated array signature projection gates
